requirements.txt
torch>=2.0.0
numpy
pillow
matplotlib
scikit-learn
networkx           # optional for future graph tools
tqdm